Separate lib functions from yarn action file
